# threadpilot/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThreadPilot AI API",
    "description": "MySQL-grounded intent workflow, incremental synchronization and read-only SQL Agent.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development; replace with your team server when deployed"
    }
  ],
  "paths": {
...
    "/api/v1/evidence/{source}/{row}": {
      "get": {
        "tags": [
          "Evidence"
        ],
        "summary": "Evidence Row",
        "description": "场景 1.3/2.1/2.2：白名单访问原始记录，展示引用对应数据。",
        "operationId": "evidence_row_api_v1_evidence__source___row__get",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source"
            }
          },
          {
            "name": "row",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Row"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evidence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/data": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "List Records",
        "description": "场景 1.2：有界分页读取最新提交的业务记录。",
        "operationId": "list_records_api_data_get",
        "parameters": [
          {
            "name": "dataset",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "orders",
                "production_log",
                "workshops"
              ],
              "type": "string",
              "default": "orders",
              "title": "Dataset"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecordRead"
                  },
                  "title": "Response List Records Api Data Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Data"
        ],
        "summary": "Create Record",
        "description": "场景 3.2：管理员确认后创建；主动重建可清除对应删除标记。",
        "operationId": "create_record_api_data_post",
        "parameters": [
          {
            "name": "x-confirm-write",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Confirm-Write"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/{record_id}": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Get Record",
        "description": "场景 1.2：刷新单条记录。",
        "operationId": "get_record_api_data__record_id__get",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Record Id"
            }
          },
          {
            "name": "dataset",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "orders",
                "production_log",
                "workshops"
              ],
              "type": "string",
              "default": "orders",
              "title": "Dataset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Data"
        ],
        "summary": "Update Record",
        "description": "场景 1.2/3.2：PUT 完整替换业务字段，要求当前版本且不允许改业务主键。",
        "operationId": "update_record_api_data__record_id__put",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Record Id"
            }
          },
          {
            "name": "x-confirm-write",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Confirm-Write"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Data"
        ],
        "summary": "Delete Record",
        "description": "场景 3.2：确认删除并记录墓碑，旧文件不会重新插入已删除记录。",
        "operationId": "delete_record_api_data__record_id__delete",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Record Id"
            }
          },
          {
            "name": "expected_version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Expected Version"
            }
          },
          {
            "name": "dataset",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "orders",
                "production_log",
                "workshops"
              ],
              "type": "string",
              "default": "orders",
              "title": "Dataset"
            }
          },
          {
            "name": "x-confirm-write",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Confirm-Write"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evidence/{dataset}/{record_id}": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Evidence",
        "description": "场景 1.3/2.2：展示查询时的当前记录、来源、时间和版本。",
        "operationId": "evidence_api_evidence__dataset___record_id__get",
        "parameters": [
          {
            "name": "dataset",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "orders",
                "production_log",
                "workshops"
              ],
              "type": "string",
              "title": "Dataset"
            }
          },
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Evidence Api Evidence  Dataset   Record Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ai/ask": {
      "post": {
        "tags": [
          "SQL Agent"
        ],
        "summary": "Ask",
        "description": "场景 1.1–2.2：独立会话不会混入原工作流的待确认操作。",
        "operationId": "ask_api_ai_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AskRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 1,
            "title": "Message"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "AskRequest"
      },
      "AskResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "business_time": {
            "type": "string",
            "title": "Business Time"
          },
          "queries": {
            "items": {
              "$ref": "#/components/schemas/QueryEvidence"
            },
            "type": "array",
            "title": "Queries"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "answer",
          "business_time",
          "queries"
        ],
        "title": "AskResponse"
      },
...
      "Evidence": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "row": {
            "type": "integer",
            "title": "Row"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "record": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Record"
          },
          "record_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "original_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "row",
          "url",
          "record"
        ],
        "title": "Evidence"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QueryEvidence": {
        "properties": {
          "sql": {
            "type": "string",
            "title": "Sql"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Rows"
          },
          "possibly_truncated": {
            "type": "boolean",
            "title": "Possibly Truncated"
          },
          "queried_at": {
            "type": "string",
            "title": "Queried At"
          }
        },
        "type": "object",
        "required": [
          "sql",
          "columns",
          "rows",
          "possibly_truncated",
          "queried_at"
        ],
        "title": "QueryEvidence"
      },
      "RecordRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "dataset": {
            "type": "string",
            "enum": [
              "orders",
              "production_log",
              "workshops"
            ],
            "title": "Dataset"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "source_row": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Row"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "dataset",
          "data",
          "version",
          "source",
          "source_row",
          "created_at",
          "updated_at"
        ],
        "title": "RecordRead"
      },
      "RecordWrite": {
        "properties": {
          "dataset": {
            "type": "string",
            "enum": [
              "orders",
              "production_log",
              "workshops"
            ],
            "title": "Dataset",
            "default": "orders"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "RecordWrite"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}